fix(web): pin the overrides Save/Reset in the DetailShell footer (PR review)
With a 40+ action catalog, in-flow actions scrolled below the list. The
editor now renders its own DetailShell with the action row in the
pinned footer slot, folding the thin wrapper into the panel component.
